+++ src/compiler/grouping/block-collector.ts
@@ -1,6 +1,6 @@
 import {DocumentVisitor} from "../visitor";
 import {Node} from "@unified-latex/unified-latex-types";
-import {visit, VisitInfo} from "@unified-latex/unified-latex-util-visit";
+import {VisitInfo} from "@unified-latex/unified-latex-util-visit";
 import {match} from "@unified-latex/unified-latex-util-match";
 import {BlockEnv} from "./block";
 import {Division} from "./division";