+++ src/compiler/grouping/figure-collector.ts
@@ -2,8 +2,6 @@ import {DocumentVisitor} from "../visitor";
 import {Node} from "@unified-latex/unified-latex-types";
 import {VisitInfo} from "@unified-latex/unified-latex-util-visit";
 import {ParserLogger} from "../logging-base";
-import {isLabelableDisplayMath} from "../metadata/util";
-import {LabeledEquation} from "./labeled-equation";
 import {match} from "@unified-latex/unified-latex-util-match";
 import {Figure} from "./figure";